+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/notifications/NotificationSettings.java
@@ -11,10 +11,7 @@ public class NotificationSettings extends JsonSettings {
   private boolean startupNotification = false;
   private boolean highscoreUpdatedNotification = true;
   private boolean highscoreCheckedNotification = true;
-
-  @JsonProperty("iScoredNotification")
-  private boolean iScoredNotification = true;
-
+  private boolean iscoredNotification = true;
   private boolean discordNotification = true;
   private boolean competitionNotification = true;
   private boolean recordingStartNotification = true;
@@ -97,12 +94,12 @@ public class NotificationSettings extends JsonSettings {
     this.highscoreCheckedNotification = highscoreCheckedNotification;
   }
 
-  public boolean isiScoredNotification() {
-    return iScoredNotification;
+  public boolean isIscoredNotification() {
+    return iscoredNotification;
   }
 
-  public void setiScoredNotification(boolean iScoredNotification) {
-    this.iScoredNotification = iScoredNotification;
+  public void setIscoredNotification(boolean iscoredNotification) {
+    this.iscoredNotification = iscoredNotification;
   }
 
   public int getDurationSec() {