+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/players/PlayerService.java
@@ -133,7 +133,7 @@ public class PlayerService {
 
     model.setDomain(player.getDomain());
     model.setName(player.getName());
-    model.setiScoredName(player.getiScoredName());
+    model.setCompetitionName(player.getCompetitionName());
     model.setInitials(player.getInitials());
     model.setAdministrative(player.isAdministrative());
     model.setManiaAccountUuid(player.getManiaAccountUuid());