+++ RELEASE_NOTES.md
@@ -25,3 +25,9 @@
 ### Bugfixes
 
 - **Highscores**: Fixed VPReg.stg file lookup.
+
+
+### Breaking Changes
+
+- Player comp. name (iscored)
+- iScored notification settings